chatbot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.5
qdrant-client==1.7.0
pydantic==2.5.0
python-dotenv==1.0.0
tiktoken==0.5.2
sentence-transformers==2.2.2
numpy==1.24.3
asyncio==3.4.3
httpx==0.25.2